+++ src/NzbDrone.Core/Metadata/Jikan/JikanService.cs
@@ -3,11 +3,9 @@ using System.Collections.Generic;
 using System.Diagnostics.CodeAnalysis;
 using System.Linq;
 using System.Net;
-using FuzzySharp;
 using Indexarr.Core.Metadata.Jikan.Resource;
 using NLog;
 using NzbDrone.Common.Cloud;
-using NzbDrone.Common.Extensions;
 using NzbDrone.Common.Http;
 using NzbDrone.Common.Serializer;
 using NzbDrone.Core.Metadata.Jikan.Result;
@@ -91,40 +89,6 @@ public class JikanService : MetadataSource, IJikanService
             {
                 return true;
             }
-        }
-
-        return false;
-    }
-
-    private bool CompareTitles(string left, string right)
-    {
-        if (left == right)
-        {
-            return true;
-        }
-
-        if (left.EqualsIgnoreCase(right))
-        {
-            return true;
-        }
-
-        var leftAlphaNumeric = new string(left.ToLower().Where(char.IsLetterOrDigit).ToArray());
-        var rightAlphaNumeric = new string(right.ToLower().Where(char.IsLetterOrDigit).ToArray());
-        if (leftAlphaNumeric.EqualsIgnoreCase(rightAlphaNumeric))
-        {
-            return true;
-        }
-
-        var ratio = Fuzz.Ratio(leftAlphaNumeric, rightAlphaNumeric);
-
-        if (ratio >= 99)
-        {
-            return true;
-        }
-
-        if (ratio >= 95)
-        {
-            _logger.Info("Not quite a match but close with a ratio of {0} for {1} and {2}", ratio, left, right);
         }
 
         return false;
