+++ src/NzbDrone.Core/Metadata/MangaUpdates/MangaUpdatesService.cs
@@ -3,7 +3,6 @@ using System.Collections.Generic;
 using System.Diagnostics.CodeAnalysis;
 using System.Linq;
 using System.Net;
-using FuzzySharp;
 using NLog;
 using NzbDrone.Common.Cloud;
 using NzbDrone.Common.Extensions;
@@ -124,40 +123,6 @@ public class MangaUpdatesService : MetadataSource, IMangaUpdatesService
             {
                 return true;
             }
-        }
-
-        return false;
-    }
-
-    private bool CompareTitles(string left, string right)
-    {
-        if (left == right)
-        {
-            return true;
-        }
-
-        if (left.EqualsIgnoreCase(right))
-        {
-            return true;
-        }
-
-        var leftAlphaNumeric = new string(left.ToLower().Where(char.IsLetterOrDigit).ToArray());
-        var rightAlphaNumeric = new string(right.ToLower().Where(char.IsLetterOrDigit).ToArray());
-        if (leftAlphaNumeric.EqualsIgnoreCase(rightAlphaNumeric))
-        {
-            return true;
-        }
-
-        var ratio = Fuzz.Ratio(leftAlphaNumeric, rightAlphaNumeric);
-
-        if (ratio >= 99)
-        {
-            return true;
-        }
-
-        if (ratio >= 95)
-        {
-            _logger.Info("Not quite a match but close with a ratio of {0} for {1} and {2}", ratio, left, right);
         }
 
         return false;
